Implemented new feature for remembering split positions
Instead of calling "splitWord" the method "splitWordIntoParts" can get
used. This method will return a list of "WordPart" objects instead of
plain strings. A "WordPart" object does still contain the original
string of which the current piece got splitted and also the start and
end offset of the piece in the original.